% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_chrome_browserhistory.R
\name{read_chrome_browser_history}
\alias{read_chrome_browser_history}
\title{Read Chrome browser history data}
\usage{
read_chrome_browser_history(path)
}
\arguments{
\item{path}{path to the json file}
}
\value{
a dataframe with the following variables:
\describe{
\item{page_transition}
\item{title}
\item{url}
\item{client_id}
\item{favicon_url}
\item{time_visited}
}
\description{
Read Chrome browser history data
}
