% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_google_play_store_installs.R
\name{read_google_play_store_installs}
\alias{read_google_play_store_installs}
\title{Read Google Play Store Installs}
\usage{
read_google_play_store_installs(path)
}
\arguments{
\item{path}{path to the json file}
}
\value{
a dataframe with the following variables:
\describe{
\item{firstInstallationTime}
\item{lastUpdateTime}
\item{doc.documentType}
\item{doc.title}
\item{deviceAttribute.model}
\item{deviceAttribute.carrier}
\item{deviceAttribute.manufacturer}
\item{deviceAttribute.carrier}
\item{deviceAttribute.deviceDisplayName}

...
}
}
\description{
Read Google Play Store Installs
}
