% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_youtube_history.R
\name{read_youtube_history}
\alias{read_youtube_history}
\title{Read youtube history data}
\usage{
read_youtube_history(path, encoding = "UTF-8")
}
\arguments{
\item{path}{path of the html file to read}

\item{encoding}{encoding to use for xml2::read_html}
}
\value{
a tibble
}
\description{
Reads youtube history data and returns a tibble with the video title and the date of watching the video.
}
