% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_youtube_history.R
\name{read_youtube_history}
\alias{read_youtube_history}
\title{Read youtube history data}
\usage{
read_youtube_history(path)
}
\arguments{
\item{path}{path of the html file to read}
}
\value{
tibble
}
\description{
Reads youtube history data and returns a tibble with the video title and the date of watching the video.
}
